+++ birdnet-go/CHANGELOG.md
@@ -1,3 +1,5 @@
+- Feat : provide mariadb information in the startup log to allow its usage
+
 ## 0.5.1-3 (23-04-2024)
 - Feat : Allow mounting of SMB and local drives to store the audio clips on an external drive